[AIRFLOW-5386] Move Google Dataproc to core
For more information check AIP-21.
