+++ airflow/gcp/example_dags/example_gcp_dataproc_create_cluster.py
@@ -16,51 +16,42 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
+
 """
-Example Airflow DAG for Google Dataproc PigOperator
+Example Airflow DAG that creates DataProc cluster.
 """
 
 import os
 import airflow
 from airflow import models
-from airflow.contrib.operators.dataproc_operator import (
-    DataProcPigOperator,
+from airflow.gcp.operators.dataproc import (
     DataprocClusterCreateOperator,
     DataprocClusterDeleteOperator
 )
 
-default_args = {"start_date": airflow.utils.dates.days_ago(1)}
-
-CLUSTER_NAME = os.environ.get('GCP_DATAPROC_CLUSTER_NAME', 'example-project')
 PROJECT_ID = os.environ.get('GCP_PROJECT_ID', 'an-id')
+CLUSTER_NAME = os.environ.get('GCP_DATAPROC_CLUSTER_NAME', 'example-project')
 REGION = os.environ.get('GCP_LOCATION', 'europe-west1')
-
+ZONE = os.environ.get('GCP_REGION', 'europe-west-1b')
 
 with models.DAG(
-    "example_gcp_dataproc_pig_operator",
-    default_args=default_args,
+    "example_gcp_dataproc_create_cluster",
+    default_args={"start_date": airflow.utils.dates.days_ago(1)},
     schedule_interval=None,
 ) as dag:
-    create_task = DataprocClusterCreateOperator(
-        task_id="create_task",
+    create_cluster = DataprocClusterCreateOperator(
+        task_id="create_cluster",
         cluster_name=CLUSTER_NAME,
         project_id=PROJECT_ID,
+        num_workers=2,
         region=REGION,
-        num_workers=2
     )
 
-    pig_task = DataProcPigOperator(
-        task_id="pig_task",
-        query="define sin HiveUDF('sin');",
-        region=REGION,
-        cluster_name=CLUSTER_NAME
-    )
-
-    delete_task = DataprocClusterDeleteOperator(
-        task_id="delete_task",
+    delete_cluster = DataprocClusterDeleteOperator(
+        task_id="delete_cluster",
         project_id=PROJECT_ID,
         cluster_name=CLUSTER_NAME,
         region=REGION
     )
 
-    create_task >> pig_task >> delete_task
+    create_cluster >> delete_cluster  # pylint: disable=pointless-statement
