+++ tests/gcp/hooks/test_dataproc.py
@@ -21,8 +21,8 @@
 import unittest
 from mock import MagicMock
 
-from airflow.contrib.hooks.gcp_dataproc_hook import _DataProcJob
-from airflow.contrib.hooks.gcp_dataproc_hook import DataProcHook
+from airflow.gcp.hooks.dataproc import _DataProcJob
+from airflow.gcp.hooks.dataproc import DataProcHook
 from tests.contrib.utils.base_gcp_mock import GCP_PROJECT_ID_HOOK_UNIT_TEST
 from tests.compat import mock
 
@@ -31,7 +31,7 @@ GCP_REGION = 'global'
 TASK_ID = 'test-task-id'
 
 BASE_STRING = 'airflow.contrib.hooks.gcp_api_base_hook.{}'
-DATAPROC_STRING = 'airflow.contrib.hooks.gcp_dataproc_hook.{}'
+DATAPROC_STRING = 'airflow.gcp.hooks.dataproc.{}'
 
 
 def mock_init(self, gcp_conn_id, delegate_to=None):  # pylint: disable=unused-argument
